requirements.txt
kociemba==1.2.1
opencv-python==4.11.0.86
pyserial==3.5
python-dotenv==1.1.1
RPi.GPIO==0.7.1
